requirements.txt
google-generativeai
profanity
python-dotenv
langchain 
py-cord